Rework Auth Bearer Token Injection
